+++ app/core/api/views.py
@@ -1,7 +1,6 @@
 import os
 
 from django.contrib.auth import get_user_model
-from django.shortcuts import render
 from drf_spectacular.types import OpenApiTypes
 from drf_spectacular.utils import OpenApiExample
 from drf_spectacular.utils import OpenApiParameter
@@ -43,30 +42,6 @@ from .serializers import SkillSerializer
 from .serializers import StackElementTypeSerializer
 from .serializers import TechnologySerializer
 from .serializers import UserSerializer
-
-
-def cognito_login(request):
-    cognito_domain = os.getenv(
-        "COGNITO_DOMAIN", "default_value"
-    )  # Replace 'default_value' with a default value or leave it empty
-    cognito_client_id = os.getenv("COGNITO_CLIENT_ID", "default_value")
-    cognito_redirect_uri = os.getenv("COGNITO_REDIRECT_URI", "default_value")
-    cognito_callback_url = os.getenv("COGNITO_CALLBACK_URL", "default_value")
-    cognito_aws_region = os.getenv("COGNITO_AWS_REGION", "default_value")
-
-    error_message = None
-    return render(
-        request,
-        "accounts/cognito_login.html",
-        {
-            "cognito_domain": cognito_domain,
-            "cognito_client_id": cognito_client_id,
-            "cognito_redirect_uri": cognito_redirect_uri,
-            "cognito_aws_region": cognito_aws_region,
-            "cognito_callback_url": cognito_callback_url,
-            "error_message": error_message,
-        },
-    )
 
 
 class UserProfileAPIView(RetrieveModelMixin, GenericAPIView):
